Extract shared types out of store.rs.
The store is getting really big and it's time to break it down into
smaller files. Given that these types have no actual functionality and
are pure data, they are easy to lift out into types.rs. This is just a
slice of what will be a much larger refactor.
