+++ src/store/mod.rs
@@ -4,14 +4,19 @@
 mod ready_index;
 mod scheduled_index;
 mod store;
+mod types;
 
 pub use store::{
-    BackoffConfig, BulkCompleteResult, BulkDeleteOptions, BulkPatchOptions, CommitMode,
-    DEFAULT_BACKOFF_BASE_MS, DEFAULT_BACKOFF_EXPONENT, DEFAULT_BACKOFF_JITTER_MS,
-    DEFAULT_CACHE_SIZE, DEFAULT_COMPLETED_RETENTION_MS, DEFAULT_DATA_TABLE_SIZE,
-    DEFAULT_DEAD_RETENTION_MS, DEFAULT_INDEX_TABLE_SIZE, DEFAULT_JOURNAL_SIZE,
-    DEFAULT_L0_THRESHOLD, DEFAULT_RETRY_LIMIT, EnqueueOptions, EnqueueResult, EnvConfigError,
-    ErrorRecord, FailureOptions, Job, JobStatus, ListErrorsOptions, ListErrorsPage,
-    ListJobsOptions, ListJobsPage, PatchJobOptions, RetentionConfig, RetentionConfigPatch,
-    ScanDirection, StorageConfig, Store, StoreError, StoreEvent, UniqueConstraint, UniqueWhile,
+    BulkCompleteResult, BulkDeleteOptions, BulkPatchOptions, CommitMode, DEFAULT_BACKOFF_BASE_MS,
+    DEFAULT_BACKOFF_EXPONENT, DEFAULT_BACKOFF_JITTER_MS, DEFAULT_CACHE_SIZE,
+    DEFAULT_COMPLETED_RETENTION_MS, DEFAULT_DATA_TABLE_SIZE, DEFAULT_DEAD_RETENTION_MS,
+    DEFAULT_INDEX_TABLE_SIZE, DEFAULT_JOURNAL_SIZE, DEFAULT_L0_THRESHOLD, DEFAULT_RETRY_LIMIT,
+    EnqueueOptions, EnqueueResult, EnvConfigError, FailureOptions, ListErrorsOptions,
+    ListErrorsPage, ListJobsOptions, ListJobsPage, PatchJobOptions, RetentionConfigPatch,
+    StorageConfig, Store, StoreEvent,
+};
+
+pub use types::{
+    BackoffConfig, ErrorRecord, Job, JobStatus, RetentionConfig, ScanDirection, StoreError,
+    UniqueConstraint, UniqueWhile,
 };
